Get basic dependency resolution working
The entire dependency resolution is outsourced to dask for now.

This means we only have to build a dictionary with one entry
per target, and dask will figure out in which order we need to
run the tasks.

We still need to incorporate exists() and newer() to make sure that
only necessary tasks are executed.