Implement timestamp-based skipping of targets
This is a very rough prototype implementation of skipping
up to date tasks. The timestamps of all targets are checked
on require(). A task needs to be updated if any of its dependencies
need to be updated, or if its timestamp is older than any of its
dependencies.

Ideally, there would be a database that would allow one to mix
timestamps and hashes to determine up to date-ness.