Hack PyTarget into the system
Add PyTarget, a target that persists and checkpoints (more or less)
arbitrary Python objects.
During the implementation, it has become apparant that more time
will have to be spent on the basic architecture of datapipe tasks and
targets.
The current implementation is based on a lot of hacks, and just serves
as a proof of concept.
It will probably have to be rewritten very dramatically.