Achieve large speedup by redesigning persistence
Previously, entire targets were persisted to the db via pickle.
But this is not strictly necessary: We are only interested in at most
a handful of objects that are needed for `is_damaged()`.

By saving these in the Target._memory dict, and persisting it with
simplejson, execution time can be decreased a lot.